requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.2
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
#matplotlib==3.4.1
mosquito==0.3.2
mysql-connector-python==8.0.24
mysqlclient==2.0.3
numpy==1.20.2
openpyxl==3.0.7
paho-mqtt==1.5.1
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
protobuf==3.15.8
psycopg2==2.8.6
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
textwrap3==0.9.2
threadpoolctl==2.1.0
#torch==1.8.1
tqdm==4.60.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
